pyproject.toml
[project]
name = "here-pheri"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.6.3",
    "duckdb>=1.3.1",
    "langchain-community>=0.3.26",
    "langchain-groq>=0.3.4",
    "langchain-tavily>=0.2.4",
    "langgraph>=0.5.0",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
    "textual>=3.5.0",
]
